Deployment Testing

+++ kollekt/__init__.py
@@ -17,8 +17,9 @@ def create_app():
 
     with app.app_context():
         from . import routes
-        db.drop_all()
-        print("Dropped All Tables")
+        if CONFIG_TYPE != 'config.ProductionConfig':
+            db.drop_all()
+            print("Dropped All Tables")
         db.create_all()
 
         print("Created All Tables From Models")
